refactor: remover pasta pages dentro de admin
