+++ src/app/features/admin/components/student-registration/student-registration.component.html
@@ -1,3 +1,4 @@
+<!-- TODO: foi criado um componente para esse botão em outra branch, utilizar <app-return-button> -->
 <a class="btn" routerLink="/administrador">
   <button mat-icon-button>
     <mat-icon>arrow_back</mat-icon>
