+++ src/app/features/admin/components/student-registration/student-registration.component.scss
